[Security Solution] [Attack discovery] Removes Alerts filtering feature flag (#209851)
### [Security Solution] [Attack discovery] Removes Alerts filtering feature flag

This PR removes the feature flag for the Attack discovery _Alerts filtering_ feature, introduced in <https://github.com/elastic/kibana/pull/205070>.

The `Attack discovery settings` flyout shown in the gif below is now available for all Attack discovery users when they click the settings gear:

![00_alerts_filtering](https://github.com/user-attachments/assets/1a81413b-b8f4-4965-a006-25fb529668a6)

#### Details

This PR _removes_ the `attackDiscoveryAlertFiltering` feature flag, which was configured via the following setting in kibana.dev.yml:

```yaml
xpack.securitySolution.enableExperimental:
  - 'attackDiscoveryAlertFiltering'
```

#### Desk testing

1. Ensure the `attackDiscoveryAlertFiltering` feature flag is present in `kibana.dev.yml`:

```yaml
xpack.securitySolution.enableExperimental:
  - 'attackDiscoveryAlertFiltering'
```

2. Start Kibana server and observe the startup logs

**Expected result**

- A warning that the `attackDiscoveryAlertFiltering` flag is no longer supported is logged:

```
[2025-02-05T11:24:31.612-05:00][WARN ][plugins.securitySolution.config] Unsupported "xpack.securitySolution.enableExperimental" values detected.
The following configuration values are no longer supported and should be removed from the kibana configuration file:

    xpack.securitySolution.enableExperimental:
      - attackDiscoveryAlertFiltering
```

3. Remove the `attackDiscoveryAlertFiltering` feature flag from `kibana.dev.yml`

4. Restart Kibana server and once again observe the startup logs

**Expected result**

- The warning displayed in step 2 is NOT logged

5. Navigate to Security > Attack discovery

6. Click the `Settings` gear

**Expected result**

- The `Attack discovery settings` flyout shown in the PR description above appears

(cherry picked from commit 42d2464157b8c8ae07e21762e34d652c9a27d377)
