+++ x-pack/platform/packages/shared/kbn-elastic-assistant-common/impl/capabilities/index.ts
@@ -21,6 +21,5 @@ export type AssistantFeatureKey = keyof AssistantFeatures;
 export const defaultAssistantFeatures = Object.freeze({
   assistantModelEvaluation: false,
   defendInsights: true,
-  attackDiscoveryAlertFiltering: false,
   contentReferencesEnabled: false,
 });
