+++ x-pack/platform/packages/shared/kbn-elastic-assistant-common/impl/schemas/capabilities/get_capabilities_route.schema.yaml
@@ -22,15 +22,12 @@ paths:
                 properties:
                   assistantModelEvaluation:
                     type: boolean
-                  attackDiscoveryAlertFiltering:
-                    type: boolean
                   contentReferencesEnabled:
                     type: boolean
                   defendInsights:
                     type: boolean
                 required:
                   - assistantModelEvaluation
-                  - attackDiscoveryAlertFiltering
                   - contentReferencesEnabled
                   - defendInsights
         '400':
