+++ x-pack/solutions/security/plugins/security_solution/public/attack_discovery/pages/index.test.tsx
@@ -9,7 +9,6 @@ import { mockCasesContext } from '@kbn/cases-plugin/public/mocks/mock_cases_cont
 import { dataViewPluginMocks } from '@kbn/data-views-plugin/public/mocks';
 import { createFilterManagerMock } from '@kbn/data-plugin/public/query/filter_manager/filter_manager.mock';
 import { createStubDataView } from '@kbn/data-views-plugin/common/data_view.stub';
-import { defaultAssistantFeatures } from '@kbn/elastic-assistant-common';
 import { UpsellingService } from '@kbn/security-solution-upselling/service';
 import { Router } from '@kbn/shared-ux-router';
 import { fireEvent, render, screen } from '@testing-library/react';
@@ -38,18 +37,6 @@ const mockConnectors: unknown[] = [
     actionTypeId: '.gen-ai',
   },
 ];
-
-jest.mock('@kbn/elastic-assistant-common', () => {
-  const original = jest.requireActual('@kbn/elastic-assistant-common');
-
-  return {
-    ...original,
-    defaultAssistantFeatures: {
-      ...original.defaultAssistantFeatures,
-      attackDiscoveryAlertFiltering: jest.mocked<boolean>(false), // <-- feature flag is off by default
-    },
-  };
-});
 
 jest.mock('react-use/lib/useLocalStorage', () => jest.fn().mockReturnValue(['test-id', jest.fn()]));
 jest.mock('react-use/lib/useSessionStorage', () =>
@@ -223,8 +210,6 @@ describe('AttackDiscovery', () => {
     });
 
     (useLocalStorage as jest.Mock).mockReturnValue(['test-id', jest.fn()]);
-
-    (defaultAssistantFeatures.attackDiscoveryAlertFiltering as jest.Mocked<boolean>) = false; // reset feature flag to off
   });
 
   describe('page layout', () => {
@@ -518,34 +503,8 @@ describe('AttackDiscovery', () => {
     });
   });
 
-  describe('when the attackDiscoveryAlertFiltering feature flag is off', () => {
+  describe('Alerts filtering feature', () => {
     beforeEach(() => {
-      render(
-        <TestProviders>
-          <Router history={historyMock}>
-            <UpsellingProvider upsellingService={mockUpselling}>
-              <AttackDiscoveryPage />
-            </UpsellingProvider>
-          </Router>
-        </TestProviders>
-      );
-    });
-
-    it('invokes fetchAttackDiscoveries with just the size parameter when the generate button is clicked', () => {
-      const generate = screen.getAllByTestId('generate');
-
-      fireEvent.click(generate[0]);
-
-      expect(
-        (useAttackDiscovery as jest.Mock)().fetchAttackDiscoveries as jest.Mock
-      ).toHaveBeenCalledWith({ size: 20 });
-    });
-  });
-
-  describe('when the attackDiscoveryAlertFiltering feature flag is on', () => {
-    beforeEach(() => {
-      (defaultAssistantFeatures.attackDiscoveryAlertFiltering as jest.Mocked<boolean>) = true; // <-- feature flag is on
-
       render(
         <TestProviders>
           <Router history={historyMock}>
