+++ x-pack/solutions/security/plugins/security_solution/public/attack_discovery/pages/settings_flyout/footer/translations.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const CANCEL = i18n.translate(
+  'xpack.securitySolution.attackDiscovery.settingsFlyout.cancelButton',
+  {
+    defaultMessage: 'Cancel',
+  }
+);
+
+export const SAVE = i18n.translate(
+  'xpack.securitySolution.attackDiscovery.settingsFlyout.saveButton',
+  {
+    defaultMessage: 'Save',
+  }
+);
+
+export const RESET = i18n.translate(
+  'xpack.securitySolution.attackDiscovery.settingsFlyout.resetLabel',
+  {
+    defaultMessage: 'Reset',
+  }
+);
